+++ semcore/ellipsis/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [16.0.4] - 2025-07-02
+## [16.0.5] - 2025-07-07
 
 ### Fixed
 
-- Ellipsis with `trim=middle` cut off text at the end.
+* Ellipsis with `trim=middle` cut off text at the end.
+
+## [16.0.4] - 2025-07-04
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/tooltip` [16.0.3 ~> 16.0.4], `@semcore/base-components` [16.1.0 ~> 16.1.1]).
 
 ## [16.0.3] - 2025-06-23